bug fixes 0000129 1 A4GL problem feature assigned (maubury) 07-04-03 Windows and special lines such as PROMPT LINE, MESSAGE LINE. 0000130 3 A4GL problem feature assigned (maubury) 07-03-03 MENU handling. 0000133 1 A4GL problem feature assigned (maubury) 07-03-03 CHAR field loses its NULL-ness in BEFORE FIELD. 0000119 3 A4GL problem feature feedback (maubury) 07-03-03 Incorrect expression evaluation. 0000108 6 A4GL problem feature feedback (maubury) 07-03-03 A4GL_form_loop wipes out data. 0000060 10 A4GL problem feature assigned (maubury) 07-02-03 INPUT ARRAY run time problems 0000126 3 A4GL problem feature assigned (maubury) 07-02-03 foreach ... into ... returns the error 400

+++ aubit4glsrc/trunk/incl/a4gl_libaubit4gl.h
@@ -24,7 +24,7 @@
 # | contact afalout@ihug.co.nz                                           |
 # +----------------------------------------------------------------------+
 #
-# $Id: a4gl_libaubit4gl.h,v 1.69 2003-06-30 17:36:14 mikeaubury Exp $
+# $Id: a4gl_libaubit4gl.h,v 1.70 2003-07-04 09:43:38 mikeaubury Exp $
 #
 */
 
@@ -1122,7 +1122,7 @@ enum cmd_types {
   /* ======================= From buildtin_d.c ==================== */
 
   void A4GL_push_dtime (struct A4GLSQL_dtime *p);
-  void A4GL_push_int (int p);
+  void A4GL_push_int (short p);
   void A4GL_push_long (long p);
   void A4GL_push_date (long p);
   void A4GL_push_float (float p);
